pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dream-research"
version = "0.1.0"
description = "A framework for analyzing LLM personality phase transitions and thermodynamic properties"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]

dependencies = [
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "openai>=0.27.0",
    "tenacity>=8.0.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=0.9",
] 